+++ model/orm_test.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"testing"
 )
-
+//TODO: 将测试分开
 var (
 	case1 = []struct {
 		name     string
