+++ model/post.go
@@ -1,8 +1,10 @@
 package model
 
 import (
-	"gorm.io/gorm"
 	"time"
+
+	"gorm.io/gorm"
+	"gorm.io/gorm/clause"
 )
 
 // Post 指代 帖子相关的内容
@@ -29,3 +31,33 @@ type Post struct {
 // 	UpdatedAt time.Time      `gorm:"index"`
 // 	DeletedAt gorm.DeletedAt `gorm:"index"`
 // }
+
+//GetPostByID will get a post by pid
+func GetPostByID(pid int) (post *Post, err error) {
+	post = &Post{}
+	result := db.Find(&post, pid)
+	if post.ID == 0 || result.RowsAffected == 0 {
+		err = ERR_POST_DOES_NOT_EXISTED
+	} else {
+		err = db.Preload("Comment.RefComment").Preload(clause.Associations).Find(post, pid).Error
+	}
+	return
+}
+//CreatePost will create a post
+func (user *User) CreatePost(pathToFile string) (*Post, error) {
+	post := &Post{
+		UserInfoID: user.UserInfo.ID,
+		TUrl:       pathToFile,
+		Status:     state_normal,
+	}
+	result := db.Create(&post)
+	err := result.Error
+	return post, err
+}
+
+//GetAllComment will get all comment belong to a post
+func (post *Post) GetAllComment()(c *[]Comment,err error){
+	err  = db.Preload(clause.Associations).Find(post).Error
+	c = &post.Comment
+	return
+	}
